edited current views for token auth

+++ api/views.py
@@ -6,6 +6,7 @@ from rest_framework.response import Response
 from django.contrib.auth.models import User
 from rest_framework import status
 from django.http import Http404
+from rest_framework.permissions import AllowAny
 
 
 
@@ -13,18 +14,18 @@ class ListCreateAds(APIView):
     def get(self, request):
         ads = Advertisement.objects.all()
         serializer = AdSerializer(ads, many=True)
-        return Response (data=serializer.data)
+        return Response(data=serializer.data)
 
     def post(self, request):
         serializer = AdSerializer(data=request.data)
         if serializer.is_valid():
-            serializer.save(user=User.objects.get(is_staff=True))
+            serializer.save(user=request.user)
             return Response (data=serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response (data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
-class RetrievUpdateDeleteAd(APIView):
+class RetrieveUpdateDeleteAd(APIView):
 
     def get_object(self, pk):
         try:
@@ -32,25 +33,34 @@ class RetrievUpdateDeleteAd(APIView):
         except Advertisement.DoesNotExist:
             raise Http404
 
+    def is_owner(self, request, object):
+        return request.user == object.user
+
     def get(self, request, pk):
         ad = self.get_object(pk)
         serializer = AdSerializer(ad)
-        return Response (serializer.data)
+        return Response(serializer.data)
 
     def put(self, request, pk):
         ad = self.get_object(pk)
-        # check user
-        serializer = AdSerializer(instance=ad, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response (serializer.data)
+        if self.is_owner(request, ad):
+            serializer = AdSerializer(instance=ad, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data)
+            else:
+                return Response(data=serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response (data=serializer.errors, status=status.HTTP_400_BAD_REQUEST) 
+            return Response({'response': "You don't have permission to edit this object."})
 
     def delete(self, request, pk):
         ad = self.get_object(pk)
-        ad.delete()
-        return Response (status=status.HTTP_204_NO_CONTENT)
+        if self.is_owner(request, ad):
+            ad.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response({'response': "You don't have permission to delete this object."})
+
 
 class DisplayIndustry(APIView):
     def get(self, request, pk):
@@ -61,15 +71,18 @@ class DisplayIndustry(APIView):
             else:
                 return Response(status=status.HTTP_204_NO_CONTENT)
 
-class ListCreateUser(APIView):
-    def get (self, request):
+class ListUsers(APIView):
+    def get(self, request):
         if request.user.is_staff:
             users = User.objects.all()
-            serializer = UserSerializer(instance=users, many=True)
-            return Response(serializer.data)
+            serializer = UserSerializer(users, many=True)
+            return Response (serializer.data)
         else:
-            return Response({'response': "You don't have permission to get this data"})
+            return Response({'response' : "You don't have permission to get this data."})
 
+
+class CreateUser(APIView):
+    permission_classes = [AllowAny]
     def post(self, request):
         serializer = UserSerializer(data=request.data)
         if serializer.is_valid():
@@ -77,6 +90,7 @@ class ListCreateUser(APIView):
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
